+++ CMakeLists.txt
@@ -35,6 +35,7 @@ set (Sources
     src/rsa/padding_schemes/oaep/oaep.cpp
     src/rsa/padding_schemes/rsa_padding.cpp
     src/rsa/prime_generation/prime_generation.cpp
+    src/rsa/rsa_key_generation/rsaKeyGen.cpp
     tools/utils.cpp
 )
