+++ tests/rsa/vectors/vectors_rsa.h
@@ -11,7 +11,7 @@
 
 #pragma once
 
-#include "rsa/rsaObjects.h"
+#include "rsa/rsa_key_generation/rsaKeyGen.h"
 
 static BigInt p = "0xbf7be0c5d1e70811470fbd96f922989f783d5f12714971c5a3e70e093de76bd1f52618572e83951cdb4f7d30dec892c5aaf8112081455f7047a1fa6599c507b1cbaf6b4be223a60dfe26e80f0b4baf6cc98d04420c18cab902e248a64aa38c9ffba7dd3e1fbae71151283681e5e1058ee5c9ab476a016083f04fe0afe2736527";
 static BigInt q = "0xc56538b3e9f3f9939aba958962d632ff86e87d43d7ffdf2073d6db61886e0f6c8abd60807022398b053ecdb7d142416a7922b576f91e4d995b762f37cd6a07affb290b6879ee045b978c9fafa81c8227effdfad787e5f53b46f2d3660d9386a491b5e4b80b5617adfaadb28f95917bac51727d51fda3994392ab95051345010d";
@@ -27,8 +27,6 @@ static std::string pt = "0x102030405060708090a0b0c0d0e0f";
 static std::string ct = "0x2eab1c8d5315c5094fcc5ed524c312f5bb30bc8516f080f3f10a8d5bb4031e316087d5a0357cd13a0064a5b53df9d9cb39b2ae0835b7c1dac78c06302fdd2a85fa488534ddde50afff6563d9803dacac3716f9005151869717f2aa7da76b5b33acf6185aebe7e2c97cc408cb9a8a7e6293445ed18af827c40fc983397f4564e03d59f0d1e1032d911e551e3bad62b01afa27dd62375d4e8faeea83ebf4a7ac5908c7bc0b95b707753f69de0720d2009c905719035138af5ad14a999226505132e1b6187e5183e01393b3bc14506c6bdcb3475f3e5fe0a250c3979196767ea32bc049e72bb536cf69df9516e412d363ac7c8a4f086d382e96ef7f95a735f2cdfd";
 
 
-
-
 static BigInt p1 = "0xeecfae81b1b9b3c908810b10a1b5600199eb9f44aef4fda493b81a9e3d84f632124ef0236e5d1e3b7e28fae7aa040a2d5b252176459d1f397541ba2a58fb6599";
 static BigInt q1 = "0xc97fb1f027f453f6341233eaaad1d9353f6c42d08866b1d05a0f2035028b9d869840b41666b42e92ea0da3b43204b5cfce3352524d0416a5a441e700af461503";
 static BigInt d1 = "0xa5dafc5341faf289c4b988db30c1cdf83f31251e0668b42784813801579641b29410b3c7998d6bc465745e5c392669d6870da2c082a939e37fdcb82ec93edac97ff3ad5950accfbc111c76f1a9529444e56aaf68c56c092cd38dc3bef5d20a939926ed4f74a13eddfbe1a1cecc4894af9428c2b7b8883fe4463a4bc85b1cb3c1";
@@ -41,7 +39,6 @@ static RSAPublicKey publicKeyVector1 = { n1, e1 };
 
 static std::string pt1 = "d436e99569fd32a7c8a05bbc90d32c49";
 static std::string ct1 = "1253e04dc0a5397bb44a7ab87e9bf2a039a33d1e996fc82a94ccd30074c95df763722017069e5268da5d1c0b4f872cf653c11df82314a67968dfeae28def04bb6d84b1c31d654a1970e5783bd6eb96a024c2ca2f4a90fe9f2ef5c9c140e5bb48da9536ad8700c84fc9130adea74e558d51a74ddf85d8b50de96838d6063e0955";
-
 
 
 static BigInt p2 = "0xecf5aecd1e5515fffacbd75a2816c6ebf49018cdfb4638e185d66a7396b6f8090f8018c7fd95cc34b857dc17f0cc6516bb1346ab4d582cadad7b4103352387b70338d084047c9d9539b6496204b3dd6ea442499207bec01f964287ff6336c3984658336846f56e46861881c10233d2176bf15a5e96ddc780bc868aa77d3ce769";
