+++ tests/rsa/vectors/vectors_rsa_key_gen.h
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2023-2024 The Gestalt Project Authors. All Rights Reserved.
+ *
+ * Licensed under the MIT License. See the file LICENSE for the full text.
+ */
+
+/*
+ * vectors_rsa_oaep.h
+ *
+ */
+
+#pragma once
+
+#include "bigInt/bigInt.h"
+
+// 2048 private key:
+static BigInt p = 
+    "0xbf7be0c5d1e70811470fbd96f922989f783d5f12714971c5a3e70e093de76bd1f52618572e83951cdb4f7d30dec892c5aaf8112081455f7"
+    "047a1fa6599c507b1cbaf6b4be223a60dfe26e80f0b4baf6cc98d04420c18cab902e248a64aa38c9ffba7dd3e1fbae71151283681e5e1058e"
+    "e5c9ab476a016083f04fe0afe2736527";
+static BigInt q = 
+    "0xc56538b3e9f3f9939aba958962d632ff86e87d43d7ffdf2073d6db61886e0f6c8abd60807022398b053ecdb7d142416a7922b576f91e4d9"
+    "95b762f37cd6a07affb290b6879ee045b978c9fafa81c8227effdfad787e5f53b46f2d3660d9386a491b5e4b80b5617adfaadb28f95917bac"
+    "51727d51fda3994392ab95051345010d";
+static BigInt d = 
+    "0x92efca032ba58fb5da7d93fed4c0430d8fbbe0d4766e4ec57a15a44f75c8450eba870507cecf34e740429321c39cf9eb4c0dec690a20623"
+    "42a6c3dcd3d17eca2cfed0d639208495e5ee03918bd084f147986ce7d99131a76c9c0f33db6ef76064131bfa8b79b1eb97fb9e00a9a865382"
+    "bc7e0b0f1c1b1138a692c3d4532778c8bda41a652a50a5677bf46873d1c28c5e5c076eada0cdd16c714664c269d3bc71ff7137c38f27719bc"
+    "99680af6757b787ee63eaa109d0c9d36581fc977d694b88451e06c681d05c007dd6ab397c79acca0c94fa1ce6170d179557705ce4ba800971"
+    "3cf9de90ead70ae8fa3010d8b1b95f3abb66e6c8ede5ee4b0881465a3db8e9";
+
+// 2048 public key:
+static BigInt n = 
+    "0x93a60a41945a4670161ea632e9644630301499e9b9da7501c204da5dcb8122bcbdd3529571b9f38519f0be9ee4419a41140ce453e9ca180"
+    "014b9e8fb27ef4d2eb5089eec2c0606d10b77b443c58516dfee254fedc11d529e8d4f9df607e8e88bdb4edae1cb18bdc6213e4766557c1dc1"
+    "ad510a44c9ce7e6943513f6012322fb269f7d47b1b41341d1d172f1b43ea903cd5c2ba102c7f56995e083ce8506482018dc8c9364a0db8479"
+    "34177a6a5cf5be2c2cfe87d246dbb0d10a5586f7db18ae14d6956dcb64f9266e43a50a81243122acb9234aa5d5a76fd2428c6b3a92f1937d0"
+    "f85c4d9f769fc22a62d9886688572a68a89aabf44cbb0e86535f301bc449fb";
+static BigInt e = 65537;
+
+// CRT Components:
+static BigInt dP = 
+    "0x4bc6ab21bdba2ec01f9e040a0dd77a0c8999c7d8d58a1a02fba4c13239142edfeca2c1b926881de74886c1652894b9388e3b131408f43a1"
+    "b85cd7cf8de0c51ca336fdd072f9ba9e819aeade0de387a5893be41741db365ee274d033103928b60f846eacfdde0a9f679666f83a244155e"
+    "0263776b9070b110257686d533c934bb";
+static BigInt dQ = 
+    "0x68f91f626e3414b16e43736a906920cdcaccb070f7d956ab476f275382de24cdb02a6dbb5fe6937b5edabb43d9c0c650fd027c463ccce4a"
+    "f7d78283c20bd16417beba3e9778e203b2528df72e68fe297a2850a74bc0c8a916f758afe6691e1b47ad8a43d4baa0905b89dceab94fabd2e"
+    "3252145b19eb7a81f8e9ce61d89185b1";
+static BigInt qInv = 
+    "0x63b9cc870049c7f0f4d19a40792e3e8770049eccb1c9ac24889300fe3847a7f905d73b51a9f9be9f0675e28dede8dc9f81bdf3e9930943a"
+    "9a47322a56d5b8c327c5dc248d2107db0e902f90bd9b40af32b36727628cabb6cb5cb9aefa3215679c95fb6f8b6e0ea00d9f4c9e8ceca58d4"
+    "6c18cf66b9699b74375ac9704f0e9cb9";
